creata cartella per three.js, separate le funzioni
